put worker stuff where it belongs always use workers to render cache if possible
